Clarify name of collection classes for provider
This moves the collection classes from `PuppetX::Rustup::Provider` to
`PuppetX::Rustup::Provider::Collection` for clarity. This will be more
important if we decide to create classes for subresource objects.
